move ViewModelFactory.kt to package utils
