+++ app/src/main/java/com/triagung/githubuser/ui/activity/FavoriteActivity.kt
@@ -11,7 +11,7 @@ import com.triagung.githubuser.databinding.ActivityFavoriteBinding
 import com.triagung.githubuser.model.UserDetail
 import com.triagung.githubuser.ui.adapter.FavoriteAdapter
 import com.triagung.githubuser.viewmodel.FavoriteViewModel
-import com.triagung.githubuser.viewmodel.ViewModelFactory
+import com.triagung.githubuser.utils.ViewModelFactory
 
 class FavoriteActivity : AppCompatActivity(), FavoriteAdapter.Listener {
 
@@ -25,9 +25,7 @@ class FavoriteActivity : AppCompatActivity(), FavoriteAdapter.Listener {
         binding = ActivityFavoriteBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        binding.toolbar.setNavigationIcon(R.drawable.ic_round_arrow_back_24)
-        setSupportActionBar(binding.toolbar)
-        supportActionBar?.setDisplayShowTitleEnabled(false)
+        setToolbar()
 
         favoriteViewModel = obtainViewModel(this)
 
@@ -50,6 +48,12 @@ class FavoriteActivity : AppCompatActivity(), FavoriteAdapter.Listener {
     private fun obtainViewModel(activity: AppCompatActivity): FavoriteViewModel {
         val factory = ViewModelFactory.getInstance(activity.application)
         return ViewModelProvider(activity, factory).get(FavoriteViewModel::class.java)
+    }
+
+    private fun setToolbar() {
+        binding.toolbar.setNavigationIcon(R.drawable.ic_round_arrow_back_24)
+        setSupportActionBar(binding.toolbar)
+        supportActionBar?.setDisplayShowTitleEnabled(false)
     }
 
     private fun setupRecyclerView() {
